+++ zcash_client_sqlite/src/testing/db.rs
@@ -32,7 +32,7 @@ use zcash_primitives::{
 use zcash_protocol::{
     consensus::BlockHeight, local_consensus::LocalNetwork, memo::Memo, ShieldedProtocol,
 };
-use zip32::{fingerprint::SeedFingerprint, DiversifierIndex};
+use zip32::DiversifierIndex;
 
 use crate::{
     error::SqliteClientError, util::testing::FixedClock, wallet::init::WalletMigrator, AccountUuid,
@@ -196,9 +196,11 @@ impl DataStoreFactory for TestDbFactory {
         if let Some(migrations) = &self.target_migrations {
             migrator
                 .init_or_migrate_to(&mut db_data, migrations)
-                .unwrap();
+                .expect("wallet migration succeeds for test setup with target migrations");
         } else {
-            migrator.init_or_migrate(&mut db_data).unwrap();
+            migrator
+                .init_or_migrate(&mut db_data)
+                .expect("wallet migration succeeds for test setup with default migrations");
         }
         Ok(TestDb::from_parts(db_data, data_file))
     }