+++ zcash_transparent/CHANGELOG.md
@@ -10,6 +10,14 @@ workspace.
 
 ## [Unreleased]
 
+## [0.4.0] - PLANNED
+
+### Changed
+- The type of `zcash_transparent::bundle::Bundle::value_balance` has changed.
+  The closure provided to this method for input retrieval can now indicate that
+  an input for the given outpoint is not available, and `value_balance` will
+  return `Ok(None)` when this is the case. 
+
 ## [0.3.0] - 2025-05-30
 
 ### Changed